requirements.txt

streamlit==1.32.0
numpy==1.26.4
pypdf==3.17.4
python-dotenv==1.0.0

# Groq FIX
flask
requests
httpx==0.24.1
groq==0.7.0

sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pandas==2.2.2
